+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 ### Added
 - Added models for generic taxonomy support (Fidesplus) [#6421](https://github.com/ethyca/fides/pull/6421)
 - Added LocationSelect component and additional utilities for displaying location [#6478](https://github.com/ethyca/fides/pull/6478)
+- Added UI for conditional manual task creation feature [#6431](https://github.com/ethyca/fides/pull/6431)
 - Added streaming attachment capabilities for access packages. [#6474](https://github.com/ethyca/fides/pull/6474)
 - Added tokenized redirect for streaming enabled access package downloads. [#6489](https://github.com/ethyca/fides/pull/6489)